+++ Sources/Protocols/Animatable/Animatable.swift
@@ -92,7 +92,24 @@ public extension Animatable where Self: UIView {
     let completion = {
       promise.animationCompleted()
     }
-    switch animation ?? animationType {
+    doAnimation(animation ?? self.animationType, configuration: configuration, completion: completion)
+  }
+
+  /**
+   `autoRunAnimation` method, should be called in layoutSubviews() method
+   */
+  func autoRunAnimation() {
+    if autoRun {
+      autoRun = false
+      animate(animationType)
+    }
+  }
+}
+
+fileprivate extension UIView {
+
+  func doAnimation(_ animationType: AnimationType, configuration: AnimationConfiguration, completion: @escaping () -> Void) {
+    switch animationType {
     case let .slide(way, direction):
       slide(way, direction: direction, configuration: configuration, completion: completion)
     case let .squeeze(way, direction):
@@ -136,21 +153,9 @@ public extension Animatable where Self: UIView {
     }
   }
 
-  /**
-   `autoRunAnimation` method, should be called in layoutSubviews() method
-   */
-  func autoRunAnimation() {
-    if autoRun {
-      autoRun = false
-      animate(animationType)
-    }
-  }
-}
-
-fileprivate extension Animatable where Self: UIView {
-
   // MARK: - Animation methods
-  func slide(_ way: AnimationType.Way, direction: AnimationType.Direction,
+  func slide(_ way: AnimationType.Way,
+             direction: AnimationType.Direction,
              configuration: AnimationConfiguration,
              completion: AnimatableCompletion? = nil) {
     let values = computeValues(way: way, direction: direction, configuration: configuration, shouldScale: false)
@@ -163,7 +168,8 @@ fileprivate extension Animatable where Self: UIView {
     }
   }
 
-  func squeeze(_ way: AnimationType.Way, direction: AnimationType.Direction,
+  func squeeze(_ way: AnimationType.Way,
+               direction: AnimationType.Direction,
                configuration: AnimationConfiguration,
                completion: AnimatableCompletion? = nil) {
     let values = computeValues(way: way, direction: direction, configuration: configuration, shouldScale: true)
@@ -175,7 +181,8 @@ fileprivate extension Animatable where Self: UIView {
     }
   }
 
-  func rotate(direction: AnimationType.RotationDirection, repeatCount: Int,
+  func rotate(direction: AnimationType.RotationDirection,
+              repeatCount: Int,
               configuration: AnimationConfiguration,
               completion: AnimatableCompletion? = nil) {
     CALayer.animate({
@@ -236,7 +243,8 @@ fileprivate extension Animatable where Self: UIView {
     }
   }
 
-  func slideFade(_ way: AnimationType.Way, direction: AnimationType.Direction,
+  func slideFade(_ way: AnimationType.Way,
+                 direction: AnimationType.Direction,
                  configuration: AnimationConfiguration,
                  completion: AnimatableCompletion? = nil) {
     let values = computeValues(way: way, direction: direction, configuration: configuration, shouldScale: false)
@@ -264,9 +272,10 @@ fileprivate extension Animatable where Self: UIView {
     }
   }
 
-    func squeezeFade(_ way: AnimationType.Way, direction: AnimationType.Direction,
-                     configuration: AnimationConfiguration,
-                     completion: AnimatableCompletion? = nil) {
+  func squeezeFade(_ way: AnimationType.Way,
+                   direction: AnimationType.Direction,
+                   configuration: AnimationConfiguration,
+                   completion: AnimatableCompletion? = nil) {
     let values = computeValues(way: way, direction: direction, configuration: configuration, shouldScale: true)
     switch way {
     case .in:
@@ -511,8 +520,12 @@ fileprivate extension Animatable where Self: UIView {
     }
   }
 
-  private func springScale(fromX: Double, fromY: Double, toX: Double, toY: Double,
-                           configuration: AnimationConfiguration, completion: AnimatableCompletion? = nil) {
+  private func springScale(fromX: Double,
+                           fromY: Double,
+                           toX: Double,
+                           toY: Double,
+                           configuration: AnimationConfiguration,
+                           completion: AnimatableCompletion? = nil) {
     transform = CGAffineTransform(scaleX: CGFloat(fromX), y: CGFloat(fromY))
     UIView.animate(
       withDuration: configuration.duration,
@@ -531,8 +544,12 @@ fileprivate extension Animatable where Self: UIView {
     )
   }
 
-  private func layerScale(fromX: Double, fromY: Double, toX: Double, toY: Double,
-                          configuration: AnimationConfiguration, completion: AnimatableCompletion? = nil) {
+  private func layerScale(fromX: Double,
+                          fromY: Double,
+                          toX: Double,
+                          toY: Double,
+                          configuration: AnimationConfiguration,
+                          completion: AnimatableCompletion? = nil) {
     CALayer.animate({
       let scaleX = CAKeyframeAnimation(keyPath: "transform.scale.x")
       scaleX.values = [fromX, toX]
@@ -553,7 +570,8 @@ fileprivate extension Animatable where Self: UIView {
   }
 
 // swiftlint:enable variable_name_min_length
-  func computeValues(way: AnimationType.Way, direction: AnimationType.Direction,
+  func computeValues(way: AnimationType.Way,
+                     direction: AnimationType.Direction,
                      configuration: AnimationConfiguration,
                      shouldScale: Bool) -> AnimationValues {
     let scale = 3 * configuration.force
@@ -631,7 +649,7 @@ fileprivate extension Animatable where Self: UIView {
                    usingSpringWithDamping: configuration.damping,
                    initialSpringVelocity: configuration.velocity,
                    options: [],
-      animations: {
+                   animations: {
         self.transform = translate
       },
       completion: { completed in
@@ -664,7 +682,7 @@ fileprivate extension Animatable where Self: UIView {
                    usingSpringWithDamping: configuration.damping,
                    initialSpringVelocity: configuration.velocity,
                    options: [],
-      animations: {
+                   animations: {
         self.transform = CGAffineTransform.identity
         self.alpha = alpha
       },
@@ -685,7 +703,7 @@ fileprivate extension Animatable where Self: UIView {
                    usingSpringWithDamping: configuration.damping,
                    initialSpringVelocity: configuration.velocity,
                    options: [],
-      animations: {
+                   animations: {
         self.transform = translateAndScale
         self.alpha = alpha
       },
@@ -703,9 +721,26 @@ fileprivate extension Animatable where Self: UIView {
 }
 // swiftlint:enable variable_name_min_length
 
+// Animations for `UIBarItem`
 public extension Animatable where Self: UIBarItem {
-  // TODO: animations for `UIBarItem`
-  public func animate(completion: AnimatableCompletion? = nil) {
+
+  public func animate(_ animation: AnimationType? = nil,
+                      duration: TimeInterval? = nil,
+                      damping: CGFloat? = nil,
+                      velocity: CGFloat? = nil,
+                      force: CGFloat? = nil,
+                      view: UIView,
+                      completion: AnimatableCompletion? = nil) {
+
+    let configuration = AnimationConfiguration(damping: damping ?? self.damping,
+                                               velocity: velocity ?? self.velocity,
+                                               duration: duration ?? self.duration,
+                                               delay: 0,
+                                               force: force ?? self.force,
+                                               timingFunction: timingFunction ?? self.timingFunction)
+    view.doAnimation(animation ?? self.animationType, configuration: configuration) {
+      completion?()
+    }
   }
 }